requirements.txt
mlflow
dagshub
dvc
fastapi
uvicorn
python-dotenv
python-box
ipykernel
datasets
transformers
langchain-community
langchain-huggingface
torch
rouge

# -e .